requirements.txt
ccxt>=4.4,<4.5
pyyaml>=6.0,<7.0
aiohttp>=3.9,<4.0
asyncpg>=0.29,<1.0
